+++ packages/checkout/widgets-lib/src/widgets/swap/SwapWidgetEvents.ts
@@ -7,7 +7,7 @@ import {
   SwapRejected,
 } from '@imtbl/checkout-widgets';
 
-export function sendSwapWidgetCloseEvent() {
+export function sendSwapWidgetCloseEvent(eventTarget: Window | EventTarget) {
   const closeWidgetEvent = new CustomEvent<WidgetEvent<any>>(
     IMTBLWidgetEvents.IMTBL_SWAP_WIDGET_EVENT,
     {
@@ -19,11 +19,11 @@ export function sendSwapWidgetCloseEvent() {
   );
   // TODO: remove once fixed
   // eslint-disable-next-line no-console
-  console.log('close widget event:', closeWidgetEvent);
-  if (window !== undefined) window.dispatchEvent(closeWidgetEvent);
+  console.log('close widget event:', eventTarget, closeWidgetEvent);
+  if (eventTarget !== undefined) eventTarget.dispatchEvent(closeWidgetEvent);
 }
 
-export const sendSwapSuccessEvent = (transactionHash: string) => {
+export const sendSwapSuccessEvent = (eventTarget: Window | EventTarget, transactionHash: string) => {
   const successEvent = new CustomEvent<WidgetEvent<SwapSuccess>>(
     IMTBLWidgetEvents.IMTBL_SWAP_WIDGET_EVENT,
     {
@@ -36,11 +36,11 @@ export const sendSwapSuccessEvent = (transactionHash: string) => {
     },
   );
   // eslint-disable-next-line no-console
-  console.log('swap success event:', successEvent);
-  if (window !== undefined) window.dispatchEvent(successEvent);
+  console.log('swap success event:', eventTarget, successEvent);
+  if (eventTarget !== undefined) eventTarget.dispatchEvent(successEvent);
 };
 
-export const sendSwapFailedEvent = (reason: string) => {
+export const sendSwapFailedEvent = (eventTarget: Window | EventTarget, reason: string) => {
   const failedEvent = new CustomEvent<WidgetEvent<SwapFailed>>(
     IMTBLWidgetEvents.IMTBL_SWAP_WIDGET_EVENT,
     {
@@ -54,11 +54,11 @@ export const sendSwapFailedEvent = (reason: string) => {
     },
   );
   // eslint-disable-next-line no-console
-  console.log('swap failed event:', failedEvent);
-  if (window !== undefined) window.dispatchEvent(failedEvent);
+  console.log('swap failed event:', eventTarget, failedEvent);
+  if (eventTarget !== undefined) eventTarget.dispatchEvent(failedEvent);
 };
 
-export const sendSwapRejectedEvent = (reason: string) => {
+export const sendSwapRejectedEvent = (eventTarget: Window | EventTarget, reason: string) => {
   const rejectedEvent = new CustomEvent<WidgetEvent<SwapRejected>>(
     IMTBLWidgetEvents.IMTBL_SWAP_WIDGET_EVENT,
     {
@@ -71,5 +71,7 @@ export const sendSwapRejectedEvent = (reason: string) => {
       },
     },
   );
-  if (window !== undefined) window.dispatchEvent(rejectedEvent);
+  // eslint-disable-next-line no-console
+  console.log('swap rejected event:', eventTarget, rejectedEvent);
+  if (eventTarget !== undefined) eventTarget.dispatchEvent(rejectedEvent);
 };